fs: rotate populated inodes
The previous patch to keep populated inodes off the shrinker LRUs had
a rare race condition that caused BUG_ON()s to trigger on inode state
at freeing time.

While this bug is being addressed, and the patch is fundamentally
rewritten to incorporate upstream feedback, put in a workaround to
protect inodes with page cache and non-resident information from being
reclaimed, in order to avoid erratic VM behavior.

This workaround has been known to cause performance problems on xfs,
where a larger pool of inodes of which only some are reclaimable
attracts undue scan pressure, which in turn drives the shrinkers into
the dirty inode throttling that xfs does in the reclaim path.

For now this appears to be the better trade-off.