fs: don't shrink populated inodes
The inode shrinker drops page cache and non-resident page info,
behavior that dates back to highmem systems (lowmem might be under
inode pressure before highmem is under cache pressure), but is
extremely disruptive in our fleet:

- hot page cache can sporadically disappear under mild pressure,
  simply because the inodes holding it aren't open and age out. Think:
  working with a git tree.

- non-resident info also gets dropped this way, which prevents the VM
  from detecting, mitigating, and reporting thrashing. This messes up
  workload performance, container isolation, health detection & load
  shedding, and safe senpai operation

This patch makes the shrinker skip inodes with cache state. It's a
workaround, the real fix is keeping populated inodes off the shrinker
LRU to begin with, which is a way more invasive change. We have not
seen any negative side effects from this patch since 5.2.

It's not upstream, because it has been known to cause performance
problems on xfs, when a large pool of inodes, of which only some are
reclaimable, attracts undue scan pressure. This drives the shrinkers
into the dirty inode throttling that xfs does in the reclaim path.

For us, it appears to be the better trade-off.

[guro: fix xarray singleton not picked up by shadow shrinker]