scftorture: Flag errors in torture-compatible manner
Summary:
scftorture: Flag errors in torture-compatible manner

This commit prints error counts on the statistics line and also adds a
"!!!" if any of the counters are non-zero.  Allocation failures are
(somewhat) forgiven, but all other errors result in a "FAILURE" print
at the end of the test.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>