scftorture: Check unexpected "switch" statement value
Summary:
scftorture: Check unexpected "switch" statement value

This commit adds a "default" case to the switch statement in
scftorture_invoke_one() which contains a WARN_ON_ONCE() and an assignment
to ->scfc_out to suppress knock-on warnings.  These knock-on warnings
could otherwise cause the user to think that there was a memory-ordering
problem in smp_call_function() instead of a bug in scftorture.c itself.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>