kdump: strip down 5.6-kdump config
Summary:
We're running low on memory again, so disable a bunch of stuff
that we'll never use the kdump kernel.  There's still a boatload of
stuff we could turn off, but this is at least the worst offenders.
With this, and the initramfs changes I have pending, we win back ~20MB

Test Plan: tested on kerneltest

Reviewers: #kernel_ops

Subscribers: 

Tasks: T70597571

Tags: 

Blame Revision: 


Differential Revision: https://phabricator.intern.facebook.com/D22731100