fb: mm: fix inode reclaim issue
Summary:
Recent changes in the inode reclaim code caused a regression in the inode
reclaim process: exceptional entries are pinning inodes and prevent them
to be released.
There are two reasons:
1) a single shadow entry is embedded into the xarray root and is never
reclaimed by the shadow shrinker, so there is a ton of inodes with
exactly 0 attached pages and 1 attached shadow entry
2) there is a circular dependency between the number of inodes (which
partially determines the size of slab memory) and the number of shadow
entries, which depends on the number of pages in lruvec/node.
If shadow entries are pinning inodes, and we have a ton of slabs,
it means the maximum number of shadow entries is also big, so we
don't reclaim shadow entries and inodes.

To fix these issues, let's do the following:
1) reclaim inodes with exactly 1 shadow entry
2) exclude slab memory from  the calculation of the maximum number of shadow
entries

This is a FB-only fix, as in upstream shdaow entries are not
pinning inodes (as now). We might end up with a different version in upstream.

Signed-off-by: Roman Gushchin <guro@fb.com>

Test Plan: manually tested on scmbackup002.ash7

Reviewers: jweiner

Subscribers: clm

Differential Revision: https://phabricator.intern.facebook.com/D22827889

Tasks: T70795048